+++ src/scripts/store/dimClearNewItems.directive.html
@@ -1,3 +1,3 @@
-<div class="clear-new-items" ng-if="$ctrl.storeService.hasNewItems && $ctrl.settings.showNewItems">
-  <button ng-click="$ctrl.storeService.clearNewItems()" title="Keyboard shortcut: X"><i class="fa fa-thumbs-up"></i> <span translate="Hotkey.ClearNewItems"></span></button>
+<div class="clear-new-items" ng-if="$ctrl.newItemsService.hasNewItems && $ctrl.settings.showNewItems">
+  <button ng-click="$ctrl.newItemsService.clearNewItems()" title="Keyboard shortcut: X"><i class="fa fa-thumbs-up"></i> <span translate="Hotkey.ClearNewItems"></span></button>
 </div>
